feat: Decouple Predicate from implementations
Fixes #30
