+++ .travis.yml
@@ -13,17 +13,17 @@ matrix:
     install:
     - rustup component add rustfmt-preview
     script:
-    - cargo fmt -- --write-mode=diff
+    - cargo fmt --all -- --write-mode=diff
   - env: RUSTFLAGS="-D warnings"
     rust: 1.27.0  # `stable`: Locking down for consistent behavior
     script:
-    - cargo check --tests
+    - cargo check --all --tests
   - env: CLIPPY
     rust: nightly-2018-07-17
     install:
       - rustup component add clippy-preview
     script:
-      - cargo clippy --all-features -- -D clippy
+      - cargo clippy --all --all-features -- -D clippy
 
 install:
 - rustc -Vv
@@ -31,9 +31,9 @@ install:
 
 script:
 - cargo check --verbose --no-default-features
-- cargo check --verbose
-- cargo test --verbose
-- cargo doc --no-deps
+- cargo check --all --verbose
+- cargo test --all --verbose
+- cargo doc --all --no-deps
 
 cache:
   cargo: true
