+++ Cargo.toml
@@ -1,6 +1,9 @@
+[workspace]
+members = ["predicates-core", "predicates-tree"]
+
 [package]
 name = "predicates"
-version = "0.5.2"
+version = "0.9.0"
 authors = ["Nick Stevens <nick@bitcurry.com>"]
 license = "MIT/Apache-2.0"
 readme = "README.md"
@@ -18,13 +21,15 @@ travis-ci = { repository = "assert-rs/predicates-rs" }
 appveyor = { repository = "assert-rs/predicates-rs" }
 
 [dependencies]
+predicates-core = { version = "0.9", path = "predicates-core" }
 difference = { version = "2.0", optional = true }
 normalize-line-endings = { version = "0.2.2", optional = true }
 regex = { version="1.0", optional = true }
 float-cmp = { version="0.4", optional = true }
-treeline = { version = "0.1", optional = true }
+
+[dev-dependencies]
+predicates-tree = { version = "0.9", path = "predicates-tree" }
 
 [features]
-default = ["difference", "regex", "float-cmp", "normalize-line-endings", "tree"]
+default = ["difference", "regex", "float-cmp", "normalize-line-endings"]
 unstable = []
-tree = ["treeline",]
