+++ predicates-core/README.md
@@ -1,0 +1,26 @@
+# predicates-core
+
+> Traits for **boolean-valued predicate functions** in Rust.
+
+[![Build Status](https://travis-ci.org/assert-rs/predicates-rs.svg?branch=master)](https://travis-ci.org/assert-rs/predicates-rs)
+[![Build status](https://ci.appveyor.com/api/projects/status/enru6k55xme867u6?svg=true)](https://ci.appveyor.com/project/epage/predicates-rs)
+[![Documentation](https://img.shields.io/badge/docs-master-blue.svg)](https://docs.rs/predicates-core)
+![License](https://img.shields.io/crates/l/predicates-core.svg)
+[![Crates.io](https://img.shields.io/crates/v/predicates-core.svg?maxAge=2592000)](https://crates.io/crates/predicates-core)
+
+[Changelog](https://github.com/assert-rs/predicates-rs/blob/master/predicates-core/CHANGELOG.md)
+
+
+## License
+
+`predicates-core` is distributed under the terms of both the MIT license and the
+Apache License (Version 2.0).
+
+See LICENSE-APACHE, and LICENSE-MIT for details.
+
+
+## Credits
+
+Big thanks to [futures-rs](https://github.com/alexcrichton/futures-rs), whose
+slick API design informed a lot of decisions made on the API design of this
+library.
