+++ predicates-core/src/core.rs
@@ -30,20 +30,3 @@ pub trait Predicate<Item: ?Sized>: reflection::PredicateReflection {
         }
     }
 }
-
-pub(crate) fn default_find_case<'a, P, Item>(
-    pred: &'a P,
-    expected: bool,
-    variable: &Item,
-) -> Option<reflection::Case<'a>>
-where
-    P: Predicate<Item>,
-    Item: ?Sized,
-{
-    let actual = pred.eval(variable);
-    if expected == actual {
-        Some(reflection::Case::new(Some(pred), actual))
-    } else {
-        None
-    }
-}
