+++ predicates-tree/CHANGELOG.md
@@ -1,0 +1,9 @@
+# Change Log
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
+and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [Unreleased] - Unreleased
+
+[Unreleased]: https://github.com/assert-rs/predicates-rs/compare/v0.5.2...HEAD
