+++ src/lib.rs
@@ -92,6 +92,8 @@
 
 #![warn(missing_docs, missing_debug_implementations)]
 
+extern crate predicates_core;
+
 #[cfg(feature = "difference")]
 extern crate difference;
 #[cfg(feature = "float-cmp")]
@@ -100,16 +102,12 @@ extern crate float_cmp;
 extern crate normalize_line_endings;
 #[cfg(feature = "regex")]
 extern crate regex;
-#[cfg(feature = "treeline")]
-extern crate treeline;
 
 pub mod prelude;
 
-mod core;
-pub use core::*;
+pub use predicates_core::*;
 mod boxed;
 pub use boxed::*;
-pub mod reflection;
 
 // core predicates
 pub mod constant;
@@ -125,5 +123,5 @@ pub mod boolean;
 pub mod float;
 pub mod path;
 pub mod str;
-#[cfg(feature = "tree")]
-pub mod tree;
+
+mod utils;
