+++ src/path/fs.rs
@@ -11,8 +11,8 @@ use std::fs;
 use std::io::{self, Read};
 use std::path;
 
-use core;
 use reflection;
+use utils;
 use Predicate;
 
 fn read_file(path: &path::Path) -> io::Result<Vec<u8>> {
@@ -25,7 +25,7 @@ fn read_file(path: &path::Path) -> io::Result<Vec<u8>> {
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct BinaryFilePredicate {
     path: path::PathBuf,
-    content: reflection::DebugAdapter<Vec<u8>>,
+    content: utils::DebugAdapter<Vec<u8>>,
 }
 
 impl BinaryFilePredicate {
@@ -66,7 +66,7 @@ impl Predicate<path::Path> for BinaryFilePredicate {
         expected: bool,
         variable: &path::Path,
     ) -> Option<reflection::Case<'a>> {
-        core::default_find_case(self, expected, variable)
+        utils::default_find_case(self, expected, variable)
     }
 }
 
@@ -76,7 +76,7 @@ impl Predicate<[u8]> for BinaryFilePredicate {
     }
 
     fn find_case<'a>(&'a self, expected: bool, variable: &[u8]) -> Option<reflection::Case<'a>> {
-        core::default_find_case(self, expected, variable)
+        utils::default_find_case(self, expected, variable)
     }
 }
 
@@ -107,7 +107,7 @@ impl fmt::Display for BinaryFilePredicate {
 /// assert_eq!(false, predicate_file.eval(Path::new("Cargo.lock")));
 /// ```
 pub fn eq_file(path: &path::Path) -> BinaryFilePredicate {
-    let content = reflection::DebugAdapter::new(read_file(path).unwrap());
+    let content = utils::DebugAdapter::new(read_file(path).unwrap());
     BinaryFilePredicate {
         path: path.to_path_buf(),
         content,
@@ -139,7 +139,7 @@ impl Predicate<path::Path> for StrFilePredicate {
         expected: bool,
         variable: &path::Path,
     ) -> Option<reflection::Case<'a>> {
-        core::default_find_case(self, expected, variable)
+        utils::default_find_case(self, expected, variable)
     }
 }
 
@@ -149,7 +149,7 @@ impl Predicate<str> for StrFilePredicate {
     }
 
     fn find_case<'a>(&'a self, expected: bool, variable: &str) -> Option<reflection::Case<'a>> {
-        core::default_find_case(self, expected, variable)
+        utils::default_find_case(self, expected, variable)
     }
 }
